[python/knowpro] Caught up with some notable changes in the TS knowPro package (#1091)
Notably the IMessageCollection and ISemanticRefCollection classes in
interfaces.py and their concrete implementations in storage.py (new).
Also added knowledge.py, lots of new tests, and ended up redoing a fair
amount of miscellaneous code.

Not everything is complete, but this has been going on for way too long
and I need to commit *something*.