[python/knowpro] Solve the "first 15 minutes" problem by switching to timezone-unaware times in stage 1 (#1358)
By feeding the question-generating LLM times in UTC that aren't marked
with a timezone, and assuming that times we receive back (which can't be
marked with a timezone) are also in UTC, we get consistent times in
time-range when-clauses, and we can answer questions like "which books
were mentioned in the first 15 minutes".

Also some cleanup of TODO.md and removed redundant `# type: inore`
directives. Of the latter, we now only have 8 left in podcast.py and one
in convindex.py (and some in tests that I didn't look at yet).