[python/knowpro] Add the simplest possible MCP server (#1362)
There is one tool/function, "answer_question", which takes a string and
returns a small struct like this:

```py
class QuestionResponse:
    success: bool
    answer: str
    time_used: int  # Milliseconds
```

It's a single file and uses the "STDIO" protocol. I have only tested it
using the `mcp dev` command (I had to do some fiddling to get it to run
the server, ended up with command = "python", args="-m
typeagent.mcp.server" .

This is mostly a bragging commit to show how "simple" it is. I had to
copy a bunch of helper code from tools/utool.py to get it working
quickly.