Add endpoint_envvar argument to AsyncEmbeddingModel (#36)
Allow passing of `endpoint_envvar` into constructor.

It must either match the envvar of the entry identified by the model_name argument in the table of predefined models, or the model_name argument must not be a key in the table.

This allows passing parameters for an entirely new model (e.g. a local model with different parameters).