Fix Connection Errors caused by subtle change to test_embeddings.py (#136)
In test_embeddings.py, test_refresh_auth() does something to its
embedding_model fixture that somehow hangs around past the end of the
test, causing Connection Errors in later test files (not in its own
file).

I've tried various ways to fix this; the cleanest way appears to be to
put back the explicit import of the embedding_model fixture.

This fixes the problem when running tests locally, but in CI it still
fails. I'm at my wits' end and plan to merge this despite the CI
failures -- it's possible that the CI failures disappear once it's in
main, and even if not, at least this fixes local test runs.